+++ src/main/java/net/silentchaos512/gear/api/item/ICoreItem.java
@@ -3,6 +3,7 @@ package net.silentchaos512.gear.api.item;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.IItemProvider;
+import net.silentchaos512.gear.api.parts.IUpgradePart;
 import net.silentchaos512.gear.api.parts.PartType;
 import net.silentchaos512.gear.api.stats.ItemStat;
 import net.silentchaos512.gear.api.stats.StatInstance;
@@ -28,6 +29,12 @@ public interface ICoreItem extends IItemProvider, IStatItem, ICustomEnchantColor
         ItemStack result = new ItemStack(this);
         GearData.writeConstructionParts(result, parts);
         GearData.recalculateStats(result, null);
+        // Allow upgrade parts to add additional data
+        parts.forEach(p -> {
+            if (p.getPart() instanceof IUpgradePart) {
+                ((IUpgradePart) p.getPart()).onAddToGear(result, p.getCraftingItem());
+            }
+        });
         // Allow traits to make any needed changes (must be done after a recalculate)
         TraitHelper.activateTraits(result, 0, (trait, level, nothing) -> {
             trait.onGearCrafted(new TraitActionContext(null, level, result));