+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Stat multiplier configs for gear. These allow the stat values of all items to be increased or decreased without overwriting every part file.
+
 ## [1.3.5] - 2019-08-06
 Updated for Forge 28.0.45
 ### Fixed